+++ tests/fixtures/FileFixture.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Hirtz\Media\video\tests\fixtures;
 
-use Hirtz\Media\models\File;
+use Hirtz\Media\Models\File;
 use yii\test\ActiveFixture;
 
 class FileFixture extends ActiveFixture